cloud/webserver/requirements.txt
google-cloud-bigquery==3.*
hypercorn<2.0
quart<2.0
